fix conversions and comparisons
conversions and comparisons are closely related to the underlying type
of storage used, so it should make sense when the storage type in the
allocator is the same and for the specific storage the arena size used
is the same. further specializations may relax/tighten these
restrictions.

moreover, currently there are no conversion capabilities between
different storage types, which will cause a compile-time error.